6 Исправлений от ментора

+++ src/main/java/ru/smoke/skill/finalproject/v1/model/Employee.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,8 +16,8 @@ public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column
-    private int id;
+    private Long id;
     @Column
-    private int balance;
+    private BigDecimal balance;
 
 }
